fix #13 : router와 model에서 중복되는 코드 제거

See merge request sw_track/class_03/web_project/team4/eladin!13

+++ src/db/models/book-model.js
@@ -3,6 +3,7 @@ import { BookSchema } from "../schemas/book-schema.js";
 
 const Book = model("books", BookSchema);
 
+//TODO refactor : unique 검증 로직 추상화할 필요가 있나?
 export class BookModel {
     //basic-user만 가능한 기능
     async findAll(){
@@ -10,22 +11,42 @@ export class BookModel {
     }
 
     async findByISBN(ISBN) {
+        if(await Book.exists({ISBN : bookInfo.ISBN}) == null){
+            throw new Error(`DB에 ${bookInfo.ISBN}는 존재하지 않습니다.`)
+        }
         return Book.findOne({ ISBN });
     }
   
     // admin만 가능한 기능
     async create(bookInfo) { //Object
+        // DB에 이미 존재하지 않으면
+        if(await Book.exists({ISBN : bookInfo.ISBN}) !== null){
+            throw new Error(`DB에 ${bookInfo.ISBN}는 존재합니다.`)
+        }
         return Book.create(bookInfo);
     }
 
     async update({ ISBN, bookInfo }) {
         //DB에 있는 bookInfo랑 유저가 수정하려는 bookInfo가 다르면
-        return Book.findOneAndUpdate({ ISBN }, bookInfo, { returnOriginal: false });
+        //이미 존재하지 않으면
+        if(await Book.exists({ISBN}) == null){
+            throw new Error(`DB에 ${bookInfo.ISBN}는 존재합니다. 고유한 ISBN을 입력해주세요.`)
+        }
+        Book.findOneAndUpdate({ ISBN }, bookInfo, { returnOriginal: false });
+        return Book.create(bookInfo);
     }
 
     async deleteByISBN(ISBN) {
+        if(await Book.exists({ISBN : bookInfo.ISBN}) == null){
+            throw new Error(`DB에 ${bookInfo.ISBN}는 존재하지 않아 삭제할 수 없습니다.`)
+        }
         await Book.deleteOne({ ISBN })
-      }
+    }
+
+    //사용하지 마세요! db 초기화 용도로 사용하는 함수입니다.
+    async dangerousDeleteAll(ISBN) {
+        await Book.deleteMany({});
+    }
   }
   
   const bookModel = new BookModel();
